backend/requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0

SQLAlchemy==2.0.47
psycopg2-binary==2.9.11

pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0

prometheus-fastapi-instrumentator 
python-dotenv==1.2.1
python-multipart==0.0.22

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-jose[cryptography]==3.5.0